fix: Add transparency support and fix AI mode image caching
Transparency Support:
- Preserve ARGB32 format throughout image processing pipeline
- Apply brightness/contrast/gamma only to RGB channels, preserve alpha
- Add checkered background pattern to visualize transparency
- Fix image adjustment processing to maintain alpha channel

AI Mode Image Caching Fix:
- Invalidate SAM hash when new image loads to prevent using cached previous image
- Add proper cache invalidation in _reset_state() method
- Ensure AI mode operates on current image instead of previous cached data
- Maintain performance with lazy loading for same-image AI operations

All tests pass, transparency now fully supported for PNG images with alpha channels
