+++ src/lazylabel/ui/main_window.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import cv2
 import numpy as np
-from PyQt6.QtCore import QModelIndex, QPointF, QRectF, Qt, QThread, QTimer, pyqtSignal
+from PyQt6.QtCore import QModelIndex, QPointF, QRectF, Qt, QTimer, pyqtSignal
 from PyQt6.QtGui import (
     QBrush,
     QColor,
@@ -53,345 +53,13 @@ from .numeric_table_widget_item import NumericTableWidgetItem
 from .photo_viewer import PhotoViewer
 from .right_panel import RightPanel
 from .widgets import StatusBar
-from .workers import ImageDiscoveryWorker, SAMUpdateWorker
-
-
-class SingleViewSAMInitWorker(QThread):
-    """Worker thread for initializing single-view SAM model in background."""
-
-    model_initialized = pyqtSignal(object)  # model_instance
-    finished = pyqtSignal()
-    error = pyqtSignal(str)
-    progress = pyqtSignal(str)  # status message
-
-    def __init__(self, model_manager, default_model_type, custom_model_path=None):
-        super().__init__()
-        self.model_manager = model_manager
-        self.default_model_type = default_model_type
-        self.custom_model_path = custom_model_path
-        self._should_stop = False
-
-    def stop(self):
-        """Stop the worker gracefully."""
-        self._should_stop = True
-
-    def run(self):
-        """Initialize SAM model in background."""
-        try:
-            if self._should_stop:
-                return
-
-            if self.custom_model_path:
-                # Load custom model
-                model_name = os.path.basename(self.custom_model_path)
-                self.progress.emit(f"Loading {model_name}...")
-
-                success = self.model_manager.load_custom_model(self.custom_model_path)
-                if not success:
-                    raise Exception(f"Failed to load custom model: {model_name}")
-
-                sam_model = self.model_manager.sam_model
-            else:
-                # Initialize the default model
-                self.progress.emit("Initializing AI model...")
-                sam_model = self.model_manager.initialize_default_model(
-                    self.default_model_type
-                )
-
-            if self._should_stop:
-                return
-
-            if sam_model and sam_model.is_loaded:
-                self.model_initialized.emit(sam_model)
-                self.progress.emit("AI model initialized")
-            else:
-                self.error.emit("Model failed to load")
-
-        except Exception as e:
-            if not self._should_stop:
-                self.error.emit(f"Failed to load AI model: {str(e)}")
-
-
-class MultiViewSAMInitWorker(QThread):
-    """Worker thread for initializing multi-view SAM models in background."""
-
-    model_initialized = pyqtSignal(int, object)  # viewer_index, model_instance
-    all_models_initialized = pyqtSignal(int)  # total_models_count
-    error = pyqtSignal(str)
-    progress = pyqtSignal(int, int)  # current, total
-
-    def __init__(self, model_manager, parent=None):
-        super().__init__(parent)
-        self.model_manager = model_manager
-        self._should_stop = False
-        self.models_created = []
-
-    def stop(self):
-        """Request the worker to stop."""
-        self._should_stop = True
-
-    def run(self):
-        """Initialize multi-view SAM models in background thread."""
-        try:
-            if self._should_stop:
-                return
-
-            # Import the required model classes
-            from ..models.sam_model import SamModel
-
-            try:
-                from ..models.sam2_model import Sam2Model
-
-                SAM2_AVAILABLE = True
-            except ImportError:
-                Sam2Model = None
-                SAM2_AVAILABLE = False
-
-            # Determine which type of model to create
-            # Get the currently selected model from the GUI
-            parent = self.parent()
-            custom_model_path = None
-            default_model_type = "vit_h"  # fallback
-
-            if parent and hasattr(parent, "control_panel"):
-                # Get the selected model path from the model selection widget
-                model_path = parent.control_panel.model_widget.get_selected_model_path()
-                if model_path:
-                    # User has selected a custom model
-                    custom_model_path = model_path
-                    # Detect model type from filename
-                    default_model_type = self.model_manager.detect_model_type(
-                        model_path
-                    )
-                else:
-                    # Using default model
-                    default_model_type = (
-                        parent.settings.default_model_type
-                        if hasattr(parent, "settings")
-                        else "vit_h"
-                    )
-
-            is_sam2 = default_model_type.startswith("sam2")
-
-            # Create model instances for all viewers - but optimize memory usage
-            config = parent._get_multi_view_config()
-            num_viewers = config["num_viewers"]
-
-            # Warn about performance implications for VIT_H in multi-view
-            if num_viewers > 2 and default_model_type == "vit_h":
-                logger.warning(
-                    f"Using vit_h model with {num_viewers} viewers may cause performance issues. Consider using vit_b for better performance."
-                )
-            for i in range(num_viewers):
-                if self._should_stop:
-                    return
-
-                self.progress.emit(i + 1, num_viewers)
-
-                try:
-                    if is_sam2 and SAM2_AVAILABLE:
-                        # Create SAM2 model instance
-                        if custom_model_path:
-                            model_instance = Sam2Model(custom_model_path)
-                        else:
-                            model_instance = Sam2Model(model_type=default_model_type)
-                    else:
-                        # Create SAM1 model instance
-                        if custom_model_path:
-                            model_instance = SamModel(
-                                model_type=default_model_type,
-                                custom_model_path=custom_model_path,
-                            )
-                        else:
-                            model_instance = SamModel(model_type=default_model_type)
-
-                    if self._should_stop:
-                        return
-
-                    if model_instance and getattr(model_instance, "is_loaded", False):
-                        self.models_created.append(model_instance)
-                        self.model_initialized.emit(i, model_instance)
-
-                        # Synchronize and clear GPU cache after each model for stability
-                        try:
-                            import torch
-
-                            if torch.cuda.is_available():
-                                torch.cuda.synchronize()  # Ensure model is fully loaded
-                                torch.cuda.empty_cache()
-                        except ImportError:
-                            pass  # PyTorch not available
-                    else:
-                        raise Exception(f"Model instance {i + 1} failed to load")
-
-                except Exception as model_error:
-                    logger.error(
-                        f"Error creating model instance {i + 1}: {model_error}"
-                    )
-                    if not self._should_stop:
-                        self.error.emit(
-                            f"Failed to create model instance {i + 1}: {model_error}"
-                        )
-                    return
-
-            if not self._should_stop:
-                self.all_models_initialized.emit(len(self.models_created))
-
-        except Exception as e:
-            if not self._should_stop:
-                self.error.emit(str(e))
-
-
-class MultiViewSAMUpdateWorker(QThread):
-    """Worker thread for updating SAM model image in multi-view mode."""
-
-    finished = pyqtSignal(int)  # viewer_index
-    error = pyqtSignal(int, str)  # viewer_index, error_message
-
-    def __init__(
-        self,
-        viewer_index,
-        model,
-        image_path,
-        operate_on_view=False,
-        current_image=None,
-        parent=None,
-    ):
-        super().__init__(parent)
-        self.viewer_index = viewer_index
-        self.model = model
-        self.image_path = image_path
-        self.operate_on_view = operate_on_view
-        self.current_image = current_image
-        self._should_stop = False
-        self.scale_factor = 1.0
-
-    def stop(self):
-        """Request the worker to stop."""
-        self._should_stop = True
-
-    def get_scale_factor(self):
-        """Get the scale factor used for image resizing."""
-        return self.scale_factor
-
-    def run(self):
-        """Update SAM model image in background thread."""
-        try:
-            if self._should_stop:
-                return
-
-            # Clear GPU cache to reduce memory pressure in multi-view mode
-            try:
-                import torch
-
-                if torch.cuda.is_available():
-                    torch.cuda.empty_cache()
-            except ImportError:
-                pass  # PyTorch not available
-
-            if self.operate_on_view and self.current_image is not None:
-                # Use the provided modified image
-                if self._should_stop:
-                    return
-
-                # Optimize image size for faster SAM processing
-                image = self.current_image
-                original_height, original_width = image.shape[:2]
-                max_size = 1024
-
-                if original_height > max_size or original_width > max_size:
-                    # Calculate scaling factor
-                    self.scale_factor = min(
-                        max_size / original_width, max_size / original_height
-                    )
-                    new_width = int(original_width * self.scale_factor)
-                    new_height = int(original_height * self.scale_factor)
-
-                    # Resize using OpenCV for speed
-                    image = cv2.resize(
-                        image, (new_width, new_height), interpolation=cv2.INTER_AREA
-                    )
-                else:
-                    self.scale_factor = 1.0
-
-                if self._should_stop:
-                    return
-
-                # Set image from numpy array
-                self.model.set_image_from_array(image)
-            else:
-                # Load original image
-                if self._should_stop:
-                    return
-
-                # Optimize image size for faster SAM processing
-                pixmap = QPixmap(self.image_path)
-                if pixmap.isNull():
-                    if not self._should_stop:
-                        self.error.emit(self.viewer_index, "Failed to load image")
-                    return
-
-                original_width = pixmap.width()
-                original_height = pixmap.height()
-                max_size = 1024
-
-                if original_width > max_size or original_height > max_size:
-                    # Calculate scaling factor
-                    self.scale_factor = min(
-                        max_size / original_width, max_size / original_height
-                    )
-
-                    # Scale down while maintaining aspect ratio
-                    scaled_pixmap = pixmap.scaled(
-                        max_size,
-                        max_size,
-                        Qt.AspectRatioMode.KeepAspectRatio,
-                        Qt.TransformationMode.SmoothTransformation,
-                    )
-
-                    # Convert to numpy array for SAM
-                    qimage = scaled_pixmap.toImage()
-                    width = qimage.width()
-                    height = qimage.height()
-                    ptr = qimage.bits()
-                    ptr.setsize(height * width * 4)
-                    arr = np.array(ptr).reshape(height, width, 4)
-                    # Convert RGBA to RGB
-                    image_array = arr[:, :, :3]
-
-                    if self._should_stop:
-                        return
-
-                    # Add CUDA synchronization for multi-model scenarios
-                    try:
-                        import torch
-
-                        if torch.cuda.is_available():
-                            torch.cuda.synchronize()
-                    except ImportError:
-                        pass
-
-                    self.model.set_image_from_array(image_array)
-                else:
-                    self.scale_factor = 1.0
-
-                    # Add CUDA synchronization for multi-model scenarios
-                    try:
-                        import torch
-
-                        if torch.cuda.is_available():
-                            torch.cuda.synchronize()
-                    except ImportError:
-                        pass
-
-                    self.model.set_image_from_path(self.image_path)
-
-            if not self._should_stop:
-                self.finished.emit(self.viewer_index)
-
-        except Exception as e:
-            if not self._should_stop:
-                self.error.emit(self.viewer_index, str(e))
+from .workers import (
+    ImageDiscoveryWorker,
+    MultiViewSAMInitWorker,
+    MultiViewSAMUpdateWorker,
+    SAMUpdateWorker,
+    SingleViewSAMInitWorker,
+)
 
 
 class PanelPopoutWindow(QDialog):
@@ -4050,6 +3718,10 @@ class MainWindow(QMainWindow):
         self.crop_mode = False
         self.crop_start_pos = None
         self.current_crop_coords = None
+
+        # Reset SAM model state - force reload for new image
+        self.current_sam_hash = None  # Invalidate SAM cache
+        self.sam_is_dirty = True  # Mark SAM as needing update
 
         # Reset AI mode state
         self.ai_click_start_pos = None
