+++ src/lazylabel/ui/workers/__init__.py
@@ -1,6 +1,15 @@
 """Worker thread classes for background operations."""
 
 from .image_discovery_worker import ImageDiscoveryWorker
+from .multi_view_sam_init_worker import MultiViewSAMInitWorker
+from .multi_view_sam_update_worker import MultiViewSAMUpdateWorker
 from .sam_update_worker import SAMUpdateWorker
+from .single_view_sam_init_worker import SingleViewSAMInitWorker
 
-__all__ = ["ImageDiscoveryWorker", "SAMUpdateWorker"]
+__all__ = [
+    "ImageDiscoveryWorker",
+    "MultiViewSAMInitWorker",
+    "MultiViewSAMUpdateWorker",
+    "SAMUpdateWorker",
+    "SingleViewSAMInitWorker",
+]
